requirements.txt
alembic==1.7.7
astroid==2.9.3
black==22.1.0
blinker==1.4
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
configparser==5.2.0
cycler==0.11.0
dash==2.2.0
dash-bootstrap-components==1.0.3
dash-core-components==2.0.0
dash-extensions==0.0.71
dash-html-components==2.0.0
dash-leaflet==0.1.23
dash-table==5.0.0
distlib==0.3.4
EditorConfig==0.12.3
filelock==3.6.0
Flask==2.0.3
Flask-Caching==1.10.1
Flask-Compress==1.11
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
fonttools==4.29.1
geobuf==1.1.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
isort==5.10.1
itsdangerous==2.1.0
Jinja2==3.0.3
jsbeautifier==1.14.0
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
Mako==1.2.0
MarkupSafe==2.1.0
matplotlib==3.5.1
mccabe==0.6.1
more-itertools==8.12.0
mypy-extensions==0.4.3
numpy==1.22.3
packaging==21.3
pandas==1.4.1
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
plotly==5.6.0
protobuf==3.19.4
psycopg2-binary==2.9.3
pylint==2.12.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.3
requests==2.27.1
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
SQLAlchemy==1.4.31
tenacity==8.0.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.8
virtualenv==20.14.1
Werkzeug==2.0.3
wrapt==1.13.3
